+++ src/components/home/section4/sub-section2/index.tsx
@@ -1,87 +1,65 @@
-import React from "react";
-import styled from "styled-components";
-import mq from "../../../../utils/mq";
+import React from 'react';
+import styled from 'styled-components';
+import mq from '../../../../utils/mq';
+import {Content} from "../../shared";
+import RequestPilot from "../../../shared/public/Navigation/RequestPilot";
 const Container = styled.section`
   margin: 0 !important;
-  display: flex;
-  justify-content: center;
-  padding: 4rem;
   position: relative;
-  width: 100%;
-  background-color: ${(props: any): any => props.theme.colors.jordyBlue};
-  color: ${(props): any => props.theme.colors.white};
-  };
-  ${mq.mobile`
-    padding: 2rem;
-  `}
-  ${mq.tablet`
-    padding: 4rem;
-  `}
+  height: 40%;
+  background-color: ${(props: any): any => props.theme.colors.blue};
 `;
 const Main = styled.main`
   display: flex;
-  flex-direction: row;
-  justify-content: space-between;
+  flex-direction: column;
   align-items: center;
   width: 1200px;
-  ${mq.mobile`
-    flex-direction: column;
-  `}
-  ${mq.tablet`
-    flex-direction: row;
-  `}
-`;
-const Div = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  padding: 1.5em;
-  width: 50%;
-  height: auto;
-  background-color: ${(props): any => props.theme.colors.blue};
-  border-radius: 0 10px 10px 10px;
-  h5 {
-    margin: 10px 0;
-    font-size: 22px;
-  }
-  h2 {
-    margin: 0;
-    font-size: 44px;
+  height: 100%;
+  h4 {
+    font-weight: 500px;
+    letter-spacing: auto;
+    font-size: 20px;
+    margin-top: 0px;
+    letter-spacing: 0.1rem;
+    color: ${props => props.theme.colors.white};
   }
   blockquote {
-    width: 90%;
+    width: 30vw;
     font-weight: 500;
     font-size: 16px;
-    text-align: left;
-    font-weight: bold;
-    margin: 0;
+    text-align: center;
+    margin-top: 0.5em;
+    margin-bottom: 0.5em;
     color: ${props => props.theme.colors.white};
   }
   ${mq.mobile`
-    width: 100%;
-    margin: 15px 0;
+    text-align: center;
+    blockquote {
+      min-width: 200px;
+      padding: 7px;
+    }
   `}
-  ${mq.tablet`
-    width: 30%;
-    margin: 0;
+  ${mq.mobileLarge`
+    blockquote {
+      min-width: 400px;
+      padding: 10px;
+    }
   `}
 `;
-const SubSection2: React.FC = () => (
-    <Container>
-      <Main>
-        <Div>
-          <h5>Countries</h5>
-          <h2>100+</h2>
-        </Div>
-        <Div>
-          <h5>Postings Per Week</h5>
-          <h2>10,000+</h2>
-        </Div>
-        <Div>
-          <h5>Companies</h5>
-          <h2>10M+</h2>
-        </Div>
-      </Main>
-    </Container>
-);
+const SubSection2: React.FC = () => {
+  return (
+      <Container>
+        <Content>
+          <Main>
+            <h4>The only intelligence platform for job postings</h4>
+            <RequestPilot inverted={true} large={false}/>
+            <blockquote>
+              Analyze millions of data-points on Job openings, titles,
+              descriptions, sectors, skill sets, technology stacks and more
+            </blockquote>
+          </Main>
+        </Content>
+      </Container>
+  );
+};
 export default SubSection2;
